requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
google-generativeai>=0.5.0
google-cloud-aiplatform>=1.49.0
cloud-sql-python-connector[pymysql]
pypdf